refactor: relocate totalOutdatedDocs scope and cast dbId to long
- Move totalOutdatedDocs variable declaration from one query method to another in SolrJQueryService.java to match its actual usage context
- Cast int to long when assigning dbId field in ContentTrackerTest to ensure type compatibility

+++ search-services/alfresco-indexing-trackers/src/test/java/org/alfresco/indexing/tracker/ContentTrackerTest.java
@@ -354,7 +354,7 @@ public class ContentTrackerTest
         for (int i = 1; i <= count; i++)
         {
             TenantDbId document = new TenantDbId();
-            document.dbId = i;
+            document.dbId = (long) i;
             document.tenant = "";
             documents.add(document);
         }
